% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_info.R
\name{events_info}
\alias{events_info}
\title{Searches time ocurrence and other information of specific events}
\usage{
events_info(data, events = c("SHOT", "GOAL", "FREE KICK", "CORNER KICK"))
}
\arguments{
\item{data}{event dataframe previously loaded}

\item{events}{one or more events to search}
}
\value{
dataframe with searched events and their specific info
}
\description{
Searches time ocurrence and other information of specific events
}
