% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pitch.R
\name{get_pitch}
\alias{get_pitch}
\title{Draws a soccer pitch as a ggplot background layer}
\usage{
get_pitch(pitch_fill = "#74a9cf", pitch_col = "black")
}
\arguments{
\item{pitch_fill}{colour used to fill the pitch}

\item{pitch_col}{colour used for lines of the pitch}
}
\value{
ggplot object
}
\description{
Draws a soccer pitch as a ggplot background layer
}
