% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_data.R
\name{get_tidy_data}
\alias{get_tidy_data}
\title{A function}
\usage{
get_tidy_data(provider = "Metrica", home_team_file, away_team_file)
}
\arguments{
\item{provider}{A character variable to set the name of the tracking data provider which defines the data format}

\item{home_team_file}{A character variable to set the filename which contains the data of the home team}

\item{away_team_file}{A character variable to set the filename which contains the data of the away team}
}
\value{
The dataframe with all tidy data joined (long format)
}
\description{
A function
}
