% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_animate.R
\name{player_animate}
\alias{player_animate}
\title{Creates 2D soccer animations}
\usage{
player_animate(
  tidy_data,
  target_player,
  ini_frame,
  end_frame,
  player_col = "#ffc300",
  method = "base",
  pitch_long = 105,
  pitch_width = 68,
  pitch_fill = "#74a9cf",
  pitch_lines_col = "lightgrey",
  home_team_col = "white",
  away_team_col = "#dd3497",
  title = "",
  subtitle = "",
  provider = "Metrica",
  show_anim = T,
  export_gif = F,
  gif_name = "animation"
)
}
\arguments{
\item{tidy_data}{the processed dataframe ready to do visualizations. It could be obtained using either get_tidy_data() or process_catapult() functions.}

\item{target_player}{player whose actions will be highlighted}

\item{ini_frame}{the frame to consider as the initial point for the animation}

\item{end_frame}{the frame to consider as the ending point for the animation}

\item{player_col}{color to highlight the target player actions}

\item{method}{four different approaches to visualize: base (default), convexhull, voronoi, delaunay}

\item{pitch_long}{long of the pitch in meters}

\item{pitch_width}{width of the pitch in meters}

\item{pitch_fill}{color used to fill the pitch}

\item{pitch_lines_col}{color used for lines of the pitch}

\item{home_team_col}{color used to fill the players of the home team}

\item{away_team_col}{color used to fill the players of the away team}

\item{title}{graph title}

\item{subtitle}{graph subtitle}

\item{provider}{set the name of the tracking data provider which defines the data format}

\item{show_anim}{should the animation be shown on the viewer? (T/F)}

\item{export_gif}{should the animation be exported as gif file? (T/F)}

\item{gif_name}{the name of the exported gif file}
}
\value{
animated gif on viewer and/or as output file
}
\description{
Creates 2D animations using soccer tracking data allowing you export it as gif file
}
