% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_plot.R
\name{player_plot}
\alias{player_plot}
\title{Creates a 2D static plot marking the selected frames of a specific player}
\usage{
player_plot(
  tidy_data,
  target_player,
  start_frames,
  end_frames = NULL,
  pitch_long = 105,
  pitch_width = 68,
  pitch_fill = "#74a9cf",
  pitch_lines_col = "lightgrey",
  action_name = "sprints",
  action_col = "purple",
  label_col = "white",
  export_png = F
)
}
\arguments{
\item{tidy_data}{the processed dataframe ready to do visualizations. It could be obtained using the function get_tidy_data().}

\item{target_player}{the player whose selected frames will be marked}

\item{start_frames}{vector with the selected frames focused on when specific actions start}

\item{end_frames}{vector with the selected frames focused on when specific actions end}

\item{pitch_long}{long of the pitch in meters}

\item{pitch_width}{width of the pitch in meters}

\item{pitch_fill}{color used to fill the pitch}

\item{pitch_lines_col}{color used for lines of the pitch}

\item{action_name}{action name(s) to be highlighted}

\item{action_col}{color used for the frames marking related to a specific action}

\item{label_col}{color used for labels showing the Time when actions start}

\item{export_png}{should the plot be exported as png file? (T/F)}
}
\value{
plot with the selected frames of a player on Plots panel and/or as png output file
}
\description{
Creates a 2D static plot marking the selected frames of a specific player
}
