% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_stats.R
\name{players_stats}
\alias{players_stats}
\title{Calculate multiple players stats based on tracking data.
It includes Minutes played, distance, average/max speed in addition to distance
splitted by different speed ranges.}
\usage{
players_stats(data, speed_ranges = c("laurie", "perf"))
}
\arguments{
\item{data}{dataframe with the already processed tracking data in long format}

\item{speed_ranges}{character to set what kind of speed ranges has to be considered:
a) "laurie": walking, jogging, running, sprinting;
b) "perf": low speed activity, high speed running, very high speed running and sprinting}
}
\value{
dataframe with the results
}
\description{
Calculate multiple players stats based on tracking data.
It includes Minutes played, distance, average/max speed in addition to distance
splitted by different speed ranges.
}
