% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/players_stats_graph.R
\name{players_stats_graph}
\alias{players_stats_graph}
\title{Create a graph to compare players stats for a specific team}
\usage{
players_stats_graph(
  data,
  team = c("home", "away"),
  var = "ranges",
  export_png = F
)
}
\arguments{
\item{data}{dataframe with the players stats}

\item{team}{team to graph "home" or "away"}

\item{var}{variable to graph, default value = "ranges" for speed ranges}

\item{export_png}{should the graph be exported as PNG file? (T/F)}
}
\value{
dataframe with the results
}
\description{
Create a graph to compare players stats for a specific team
}
