% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_catapult.R
\name{process_catapult}
\alias{process_catapult}
\title{Preprocess Catapult data}
\usage{
process_catapult(
  home_data_file,
  away_data_file,
  ref_p1 = NULL,
  ref_p2 = NULL,
  ref_p3 = NULL,
  ref_p4 = NULL
)
}
\arguments{
\item{home_data_file}{set the data file path of the Home team}

\item{away_data_file}{set the data file path of the Away team}

\item{ref_pX}{c(LAT, LONG) coordinates for the specific field corners X}
}
\value{
The processed dataframe ready to do visualizations
}
\description{
Reads Catapult data on CSV (already on long format) and process it.
}
