% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer_plot_poss.R
\name{soccer_plot_poss}
\alias{soccer_plot_poss}
\title{Creates a 2D team visualization on a team level with avg. positioning splitting by ON/OFF possesion}
\usage{
soccer_plot_poss(
  tidy_data,
  events_path,
  team = c("Home", "Away"),
  pitch_long = 105,
  pitch_width = 68,
  pitch_fill = "#74a9cf",
  pitch_lines_col = "lightgrey",
  target_team_col = "#7209b7",
  opp_team_col = "#06d6a0",
  on_ball_col = "white",
  off_ball_col = "#dd3497",
  export_png = F,
  png_name = "plot",
  annotation = "NA"
)
}
\arguments{
\item{tidy_data}{the processed dataframe ready to do visualizations. It could be obtained using either get_tidy_data() or process_catapult() functions.}

\item{events_path}{path where the CSV file with eventing data is located}

\item{team}{target team}

\item{pitch_long}{long of the pitch in meters}

\item{pitch_width}{width of the pitch in meters}

\item{pitch_fill}{colour used to fill the pitch}

\item{pitch_lines_col}{colour used for lines of the pitch}

\item{target_team_col}{colour used to fill the border of players circles of the target team (top position on the graph)}

\item{opp_team_col}{colour used to fill the border of players circles of the opponent team (bottom position on the graph)}

\item{on_ball_col}{colour used to fill the players when their team was ON possession}

\item{off_ball_col}{colour used to fill the players when their team was OFF possession}

\item{export_png}{should the plot be exported as png file? (T/F)}

\item{png_name}{the name of the exported gif file}

\item{annotation}{'NA' by default. If "depth" is setted, vertical lines with lowest defender and highest forward will be added}
}
\value{
soccer plot on 'Plots' panel and/or as output PNG file
}
\description{
Creates a 2D team visualization on a team level with avg. positioning splitting by ON/OFF possesion
}
