% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccerAnimate.R
\name{soccerAnimate}
\alias{soccerAnimate}
\title{Creates 2D animations using soccer tracking data}
\usage{
soccerAnimate(
  tidy_data,
  ini_time,
  end_time,
  method = c("base", "convexhull", "voronoi"),
  pitch_fill = "#74a9cf",
  home_team_col = "white",
  away_team_col = "#dd3497",
  title = "",
  subtitle = "",
  provider = "Metrica",
  show_anim = T,
  export_gif = T,
  gifname = "output"
)
}
\arguments{
\item{tidy_data}{dataframe got it from get_tidy_data() function with soccer tracking data ready to animate}

\item{ini_time}{the time of tracking data to consider as initial time of the animation}

\item{end_time}{the time of tracking data to consider as ending time of the animation}

\item{method}{three different approaches to visualize: base, convexhull & voronoi}

\item{pitch_fill}{colour used to fill the pitch}

\item{home_team_col}{colour used to fill the players of the home team}

\item{away_team_col}{colour used to fill the players of the away team}

\item{title}{graph title}

\item{subtitle}{graph subtitle}

\item{provider}{set the name of the tracking data provider which defines the data format}

\item{show_anim}{should the animation be shown on the viewer? (T/F)}

\item{export_gif}{should the animation be exported as gif file? (T/F)}

\item{gifname}{the name of the exported gif file}
}
\value{
animated gif on viewer and/or as output file
}
\description{
Creates 2D animations using soccer tracking data
}
