% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprints_info.R
\name{sprints_info}
\alias{sprints_info}
\title{Get the start/end time of the sprints that every player made}
\usage{
sprints_info(data, team = c("home", "away"))
}
\arguments{
\item{data}{dataframe with the already processed tracking data in long format}

\item{team}{team to graph "home" or "away"}
}
\value{
dataframe with the results
}
\description{
Get the start/end time of the sprints that every player made
}
