% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_data.R
\name{get_tidy_data}
\alias{get_tidy_data}
\title{Prepares data}
\usage{
get_tidy_data(
  home_team_file,
  away_team_file,
  provider = "Metrica",
  convert_coord = T,
  pitch_long = 105,
  pitch_width = 68
)
}
\arguments{
\item{home_team_file}{set the filename which contains the data of the home team}

\item{away_team_file}{set the filename which contains the data of the away team}

\item{provider}{set the name of the tracking data provider which defines the data format}

\item{convert_coord}{A boolean variable to set if the coordinates should to be converted to meters}

\item{pitch_long}{long of the pitch in meters}

\item{pitch_width}{width of the pitch in meters}
}
\value{
The dataframe with all tidy data joined (long format)
}
\description{
Reads and converts both the home and away team rawdata to tidy data (long format) and join them.
Currently only Metrica Sports data format is supported
}
